refactoring route register hander approach
